fix mac app
