+++ apple/OmnivoreKit/Sources/Services/InternalModels/InternalNewsletterEmail.swift
@@ -56,20 +56,3 @@ extension Sequence where Element == InternalNewsletterEmail {
     return result
   }
 }
-
-extension NewsletterEmail {
-  static func lookup(byID emailID: String, inContext context: NSManagedObjectContext) -> NewsletterEmail? {
-    let fetchRequest: NSFetchRequest<Models.NewsletterEmail> = NewsletterEmail.fetchRequest()
-    fetchRequest.predicate = NSPredicate(
-      format: "%K == %@", #keyPath(NewsletterEmail.emailId), emailID
-    )
-
-    var email: NewsletterEmail?
-
-    context.performAndWait {
-      email = (try? context.fetch(fetchRequest))?.first
-    }
-
-    return email
-  }
-}
